# packages/services/apiSpec.yml
openapi: 3.0.0
info:
  title: Tegon
  description: AI First Project Management Tool
  version: 0.3.0
servers:
  - url: https://app.tegon.ai
paths:
...
  /api/v1/issue_comments:
    post:
      summary: Create an issue comment
      parameters:
        - name: issueId
          in: query
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueCommentDto'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueComment'
  /api/v1/issue_comments/linked_comment:
    post:
      summary: Create a linked comment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLinkedIssueCommentDto'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkedComment'
  /api/v1/issue_comments/linked_comment/:
    get:
      summary: Get linked comment
      parameters:
        - in: query
          name: sourceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkedComment'
  /api/v1/issue_comments/{issueCommentId}:
    get:
      summary: Get issue comment
      parameters:
        - in: path
          name: issueCommentId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueComment'

    post:
      summary: Update issue comment
      parameters:
        - in: path
          name: issueCommentId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIssueCommentDto'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueComment'
...
  /api/v1/issues/{issueId}:
    post:
      summary: Update issue
      parameters:
        - in: path
          name: issueId
          required: true
          schema:
            type: string
        - in: query
          name: teamId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIssueDto'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
    delete:
      summary: Delete issue
      parameters:
        - in: path
          name: issueId
          required: true
          schema:
            type: string
        - in: query
          name: teamId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
...
components:
  schemas:
...
    CreateIssueCommentDto:
      type: object
      properties:
        body:
          type: string
        parentId:
          type: string
        linkCommentMetadata:
          type: object
        sourceMetadata:
          type: object
    IssueComment:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deleted:
          type: string
          format: date-time
          nullable: true
        body:
          type: string
        userId:
          type: string
          nullable: true
        updatedById:
          type: string
        sourceMetadata:
          type: object
          nullable: true
        reactionsData:
          type: array
          items:
            type: object
        issueId:
          type: string
        parentId:
          type: string
          nullable: true
        attachments:
          type: array
          items:
            type: string
    Issue:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deleted:
          type: string
          format: date-time
          nullable: true
        title:
          type: string
        number:
          type: integer
        description:
          type: string
          nullable: true
        priority:
          type: integer
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        sortOrder:
          type: integer
          nullable: true
        subIssueSortOrder:
          type: integer
          nullable: true
        estimate:
          type: integer
          nullable: true
        sourceMetadata:
          type: object
          nullable: true
        isBidirectional:
          type: boolean
          nullable: true
        teamId:
          type: string
        createdById:
          type: string
          nullable: true
        updatedById:
          type: string
        subscriberIds:
          type: array
          items:
            type: string
        assigneeId:
          type: string
          nullable: true
        labelIds:
          type: array
          items:
            type: string
        stateId:
          type: string
        parentId:
          type: string
          nullable: true
        attachments:
          type: array
          items:
            type: string
        issueSuggestionId:
          type: string
          nullable: true
    LinkedComment:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deleted:
          type: string
          format: date-time
          nullable: true
        url:
          type: string
        sourceId:
          type: string
        sourceData:
          type: object
          nullable: true
        createdById:
          type: string
          nullable: true
        commentId:
          type: string
    CreateLinkedIssueCommentDto:
      type: object
      properties:
        url:
          type: string
        sourceId:
          type: string
        commentId:
          type: string
        sourceData:
          type: object
...
    UpdateIssueCommentDto:
      type: object
      properties:
        body:
          type: string
        parentId:
          type: string
        linkCommentMetadata:
          type: object
        sourceMetadata:
          type: object
...
    CreateLinkedIssueDto:
      type: object
      properties:
        url:
          type: string
        sourceId:
          type: string
        sourceData:
          type: object
          additionalProperties:
            type: string
    CreateIssueRelationDto:
      type: object
      required:
        - type
        - issueId
        - relatedIssueId
      properties:
        type:
          type: string
        issueId:
          type: string
        relatedIssueId:
          type: string
...
    UpdateIssueDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: number
        dueDate:
          type: string
        sortOrder:
          type: number
        subIssueSortOrder:
          type: number
        estimate:
          type: number
        labelIds:
          type: array
          items:
            type: string
        assigneeId:
          type: string
        stateId:
          type: string
        parentId:
          type: string
        isBidirectional:
          type: boolean
        subscriberIds:
          type: array
          items:
            type: string
        issueRelation:
          $ref: '#/components/schemas/CreateIssueRelationDto'
        attachments:
          type: array
          items:
            type: string
        userId:
          type: string
        linkIssueData:
          $ref: '#/components/schemas/CreateLinkedIssueDto'
        sourceMetadata:
          type: object
          additionalProperties:
            type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
security:
  - BearerAuth: []